+++ src/panels/config/backup/components/config/ha-backup-config-encryption-key.ts
@@ -1,13 +1,13 @@
 import { mdiDownload } from "@mdi/js";
 import { css, html, LitElement } from "lit";
 import { customElement, property, state } from "lit/decorators";
-import { fireEvent } from "../../../../common/dom/fire_event";
-import "../../../../components/ha-md-list";
-import "../../../../components/ha-md-list-item";
-import type { HomeAssistant } from "../../../../types";
-import { showChangeBackupEncryptionKeyDialog } from "../dialogs/show-dialog-change-backup-encryption-key";
-import { fileDownload } from "../../../../util/file_download";
-import { showSetBackupEncryptionKeyDialog } from "../dialogs/show-dialog-set-backup-encryption-key";
+import { fireEvent } from "../../../../../common/dom/fire_event";
+import "../../../../../components/ha-md-list";
+import "../../../../../components/ha-md-list-item";
+import type { HomeAssistant } from "../../../../../types";
+import { showChangeBackupEncryptionKeyDialog } from "../../dialogs/show-dialog-change-backup-encryption-key";
+import { fileDownload } from "../../../../../util/file_download";
+import { showSetBackupEncryptionKeyDialog } from "../../dialogs/show-dialog-set-backup-encryption-key";
 
 @customElement("ha-backup-config-encryption-key")
 class HaBackupConfigEncryptionKey extends LitElement {
