+++ src/panels/config/backup/components/config/ha-backup-config-schedule.ts
@@ -2,19 +2,19 @@ import type { PropertyValues } from "lit";
 import { css, html, LitElement, nothing } from "lit";
 import { customElement, property, state } from "lit/decorators";
 import memoizeOne from "memoize-one";
-import { fireEvent } from "../../../../common/dom/fire_event";
-import type { HaCheckbox } from "../../../../components/ha-checkbox";
-import "../../../../components/ha-md-list";
-import "../../../../components/ha-md-list-item";
-import "../../../../components/ha-md-select";
-import "../../../../components/ha-md-textfield";
-import type { HaMdSelect } from "../../../../components/ha-md-select";
-import "../../../../components/ha-md-select-option";
-import "../../../../components/ha-switch";
-import type { BackupConfig } from "../../../../data/backup";
-import { BackupScheduleState } from "../../../../data/backup";
-import type { HomeAssistant } from "../../../../types";
-import { clamp } from "../../../../common/number/clamp";
+import { fireEvent } from "../../../../../common/dom/fire_event";
+import type { HaCheckbox } from "../../../../../components/ha-checkbox";
+import "../../../../../components/ha-md-list";
+import "../../../../../components/ha-md-list-item";
+import "../../../../../components/ha-md-select";
+import "../../../../../components/ha-md-textfield";
+import type { HaMdSelect } from "../../../../../components/ha-md-select";
+import "../../../../../components/ha-md-select-option";
+import "../../../../../components/ha-switch";
+import type { BackupConfig } from "../../../../../data/backup";
+import { BackupScheduleState } from "../../../../../data/backup";
+import type { HomeAssistant } from "../../../../../types";
+import { clamp } from "../../../../../common/number/clamp";
 
 export type BackupConfigSchedule = Pick<BackupConfig, "schedule" | "retention">;
